split pk filters for sharding
