+++ pkg/vm/engine/disttae/txn_table.go
@@ -1672,17 +1672,36 @@ func (tbl *txnTable) NewReader(
 		return nil, err
 	}
 
-	pkFilters := ConstructPKFilters(tbl.tableDef, tbl.db.databaseName, ts, state, expr, tbl.proc.Load(), txn.engine.packerPool)
+	proc := tbl.proc.Load()
+
+	baseFilter := newBasePKFilter(
+		expr,
+		tbl.tableDef,
+		proc,
+	)
+
+	memFilter := newMemPKFilter(
+		tbl.tableDef,
+		ts,
+		state,
+		txn.engine.packerPool,
+		baseFilter,
+	)
+
+	blockFilter := newBlockReadPKFilter(
+		tbl.tableDef.Pkey.PkeyColName,
+		baseFilter,
+	)
 
 	blkArray := objectio.BlockInfoSlice(ranges)
-	if !pkFilters.inMemPKFilter.isValid || plan2.IsFalseExpr(expr) {
+	if !memFilter.isValid || plan2.IsFalseExpr(expr) {
 		return []engine.Reader{new(emptyReader)}, nil
 	}
 	if blkArray.Len() == 0 {
-		return tbl.newMergeReader(ctx, num, expr, pkFilters, nil, txnOffset)
+		return tbl.newMergeReader(ctx, num, expr, memFilter, blockFilter, nil, txnOffset)
 	}
 	if blkArray.Len() == 1 && engine.IsMemtable(blkArray.GetBytes(0)) {
-		return tbl.newMergeReader(ctx, num, expr, pkFilters, nil, txnOffset)
+		return tbl.newMergeReader(ctx, num, expr, memFilter, blockFilter, nil, txnOffset)
 	}
 	if blkArray.Len() > 1 && engine.IsMemtable(blkArray.GetBytes(0)) {
 		rds := make([]engine.Reader, num)
@@ -1699,7 +1718,7 @@ func (tbl *txnTable) NewReader(
 			}
 			dirtyBlks = append(dirtyBlks, blkInfo)
 		}
-		rds0, err := tbl.newMergeReader(ctx, num, expr, pkFilters, dirtyBlks, txnOffset)
+		rds0, err := tbl.newMergeReader(ctx, num, expr, memFilter, blockFilter, dirtyBlks, txnOffset)
 		if err != nil {
 			return nil, err
 		}
@@ -1708,7 +1727,7 @@ func (tbl *txnTable) NewReader(
 		}
 
 		if len(cleanBlks) > 0 {
-			rds0, err = tbl.newBlockReader(ctx, num, expr, pkFilters.blockReadPKFilter, cleanBlks, tbl.proc.Load(), orderedScan)
+			rds0, err = tbl.newBlockReader(ctx, num, expr, blockFilter, cleanBlks, tbl.proc.Load(), orderedScan)
 			if err != nil {
 				return nil, err
 			}
@@ -1726,14 +1745,15 @@ func (tbl *txnTable) NewReader(
 	for i := 0; i < blkArray.Len(); i++ {
 		blkInfos = append(blkInfos, blkArray.Get(i))
 	}
-	return tbl.newBlockReader(ctx, num, expr, pkFilters.blockReadPKFilter, blkInfos, tbl.proc.Load(), orderedScan)
+	return tbl.newBlockReader(ctx, num, expr, blockFilter, blkInfos, tbl.proc.Load(), orderedScan)
 }
 
 func (tbl *txnTable) newMergeReader(
 	ctx context.Context,
 	num int,
 	expr *plan.Expr,
-	pkFilter PKFilters,
+	memFilter memPKFilter,
+	blockFilter blockio.BlockReadFilter,
 	dirtyBlks []*objectio.BlockInfo,
 	txnOffset int, // Transaction writes offset used to specify the starting position for reading data.
 ) ([]engine.Reader, error) {
@@ -1742,7 +1762,8 @@ func (tbl *txnTable) newMergeReader(
 	rds0, err := tbl.newReader(
 		ctx,
 		num,
-		pkFilter,
+		memFilter,
+		blockFilter,
 		expr,
 		dirtyBlks,
 		txnOffset)
@@ -1825,7 +1846,8 @@ func (tbl *txnTable) newBlockReader(
 func (tbl *txnTable) newReader(
 	ctx context.Context,
 	readerNumber int,
-	pkFilter PKFilters,
+	memFilter memPKFilter,
+	blockFilter blockio.BlockReadFilter,
 	expr *plan.Expr,
 	dirtyBlks []*objectio.BlockInfo,
 	txnOffset int, // Transaction writes offset used to specify the starting position for reading data.
@@ -1855,14 +1877,12 @@ func (tbl *txnTable) newReader(
 	partReader := &PartitionReader{
 		table:     tbl,
 		txnOffset: txnOffset,
-		iter:      pkFilter.inMemPKFilter.iter,
+		iter:      memFilter.iter,
 		seqnumMp:  seqnumMp,
 		typsMap:   mp,
 	}
 
-	//tbl.Lock()
 	proc := tbl.proc.Load()
-	//tbl.Unlock()
 
 	readers[0] = partReader
 
@@ -1873,7 +1893,8 @@ func (tbl *txnTable) newReader(
 				newBlockMergeReader(
 					ctx,
 					tbl,
-					pkFilter,
+					memFilter,
+					blockFilter,
 					ts,
 					[]*objectio.BlockInfo{dirtyBlks[i]},
 					expr,
@@ -1891,7 +1912,8 @@ func (tbl *txnTable) newReader(
 			readers[i+1] = newBlockMergeReader(
 				ctx,
 				tbl,
-				pkFilter,
+				memFilter,
+				blockFilter,
 				ts,
 				[]*objectio.BlockInfo{dirtyBlks[i]},
 				expr,
@@ -1913,7 +1935,7 @@ func (tbl *txnTable) newReader(
 		ts,
 		readerNumber-1,
 		expr,
-		pkFilter.blockReadPKFilter,
+		blockFilter,
 		proc)
 	objInfos, steps := groupBlocksToObjects(dirtyBlks, readerNumber-1)
 	blockReaders = distributeBlocksToBlockReaders(
@@ -1927,7 +1949,7 @@ func (tbl *txnTable) newReader(
 			blockReader: blockReaders[i],
 			table:       tbl,
 			txnOffset:   txnOffset,
-			pkFilter:    pkFilter.inMemPKFilter,
+			pkFilter:    memFilter,
 			deletaLocs:  make(map[string][]objectio.Location),
 		}
 		readers[i+1] = bmr
@@ -2139,8 +2161,8 @@ func (tbl *txnTable) PKPersistedBetween(
 			bytes,
 			false)
 
-		basePKFilter := constructBasePKFilter(inExpr, tbl.tableDef, tbl.proc.Load())
-		blockReadPKFilter := constructBlockReadPKFilter(tbl.tableDef.Pkey.PkeyColName, basePKFilter)
+		basePKFilter := newBasePKFilter(inExpr, tbl.tableDef, tbl.proc.Load())
+		blockReadPKFilter := newBlockReadPKFilter(tbl.tableDef.Pkey.PkeyColName, basePKFilter)
 
 		return blockReadPKFilter.SortedSearchFunc
 	}
